+++ code/src/sixsq/nuvla/server/resources/nuvlabox/status_utils.clj
@@ -50,3 +50,17 @@
                                  is-nuvlabox? (assoc :online true))]
     (set-nuvlabox-online updated-resource)
     updated-resource))
+
+
+(defn set-inferred-location
+  [{:keys [parent] :as resource} inferred-location]
+  (when (some? inferred-location)
+    (try
+      (-> (crud/retrieve-by-id-as-admin parent)
+          (u/update-timestamps)
+          (assoc :inferred-location inferred-location)
+          (db/edit {:nuvla/authn auth/internal-identity}))
+      (catch Exception ex
+        (log/info parent "update inferred-location attribute failed!" ex))))
+  resource)
+